Added "memetcg" Project
memetcg is a meme based trading card card that I designed to be used by a "discord bot". Discord is a popular chat-based communication system that is used heavily by myself and peers. This bot acts as a user in the server and can be passed commands through the discord chat interface. Inventories of players are saved in a SQL database. Chatting in the server rewards the users credits which they can then spend on opening virtual packs of cards with popular memes. These cards can then be traded or used in duels.